bubble_sort

+++ bubble_sort/main.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 // Bubble Sort
-int *bubbleSort(int arr[], int number_of_elements) 
+int long *bubbleSort(int long arr[], int long number_of_elements) 
 {
     for (int i = 0; i < number_of_elements - 1; i++) 
     {
@@ -14,7 +14,7 @@ int *bubbleSort(int arr[], int number_of_elements)
             {
                 // If number is greater than j 
                 // swap and continue sorting
-                int temp = arr[j];
+                int long temp = arr[j];
                 arr[j] = arr[j + 1];
                 arr[j + 1] = temp;
             }
@@ -25,7 +25,7 @@ int *bubbleSort(int arr[], int number_of_elements)
 
 
 // Function to print array. Array is called by reference
-void print_array(int arr[], int number_of_elements) 
+void print_array(int long arr[], int long number_of_elements) 
 {
     for (int i = 0; i < number_of_elements; i++) 
         cout << arr[i] << " ";
@@ -37,15 +37,15 @@ void print_array(int arr[], int number_of_elements)
 int main() 
 {
     // Input for # of elements in the array
-    int number_of_elements;
+    int long number_of_elements;
     cin >> number_of_elements;
 
     // Input for elements of the array to be sorted
-    int arr [number_of_elements];
+    int long arr [number_of_elements];
     for (int i = 0; i < number_of_elements; i++)
         cin >> arr[i];
 
-    int *ans = bubbleSort(arr, number_of_elements);
+    int long *ans = bubbleSort(arr, number_of_elements);
     print_array(ans, number_of_elements);
 
     return 0;
